samples/python/requirements.txt
Flask==2.1.2
mysql_connector_repackaged==0.3.1
requests==2.22.0
